Trait yap::IntoTokens

pub trait IntoTokens<Item> {
    type Tokens: Tokens<Item = Item>;

    fn into_tokens(self) -> Self::Tokens;
}

A trait that is implemented by anything which can be converted into an object implementing the Tokens trait.

Required Associated Types

type Tokens: Tokens<Item = Item>

Required Methods

fn into_tokens(self) -> Self::Tokens